requirements.txt
beautifulsoup4==4.12.3
playwright==1.42.0
python-dotenv==1.0.1
requests==2.31.0
openai==1.13.3
tqdm==4.66.2 